poprawki7

+++ ServerApplication/src/libs/controllers/CRetrievePhotoTransaction.cpp
@@ -40,8 +40,8 @@ void CRetrievePhotoTransaction::RetrieveData(Photo aPhoto) {
 
         return;
     }
-//size_t
-		int vBuffSize {aPhoto.datasize.value()};
+
+    size_t vBuffSize {aPhoto.datasize.value()};
 		u8_t *vBuffer {new u8_t[vBuffSize]};
 
     vBlob.getData(vBuffer, vBuffSize);
